Deduplicate some scripts.
