+++ src-script/script-util.js
@@ -1,0 +1,41 @@
+/**
+ *
+ *    Copyright (c) 2020 Silicon Labs
+ *
+ *    Licensed under the Apache License, Version 2.0 (the "License");
+ *    you may not use this file except in compliance with the License.
+ *    You may obtain a copy of the License at
+ *
+ *        http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *    Unless required by applicable law or agreed to in writing, software
+ *    distributed under the License is distributed on an "AS IS" BASIS,
+ *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *    See the License for the specific language governing permissions and
+ *    limitations under the License.
+ */
+const { spawn } = require('child_process')
+
+// Utilities shared by scripts.
+
+function executeCmd(ctx, cmd, args) {
+  return new Promise((resolve, reject) => {
+    console.log(`🚀 Executing: ${cmd} ${args.join(' ')}`)
+    var c = spawn(cmd, args)
+    c.on('exit', (code) => {
+      if (code == 0) resolve(ctx)
+      else {
+        console.log(`👎 Program ${cmd} exited with error code: ${code}`)
+        reject()
+      }
+    })
+    c.stdout.on('data', (data) => {
+      process.stdout.write(data)
+    })
+    c.stderr.on('data', (data) => {
+      process.stderr.write('⇝ ' + data)
+    })
+  })
+}
+
+exports.executeCmd = executeCmd
